test(fix): fix authorizer test

+++ src/main/resources/db/migration/V43_41__Create_community_authorization_table.sql
@@ -2,5 +2,5 @@ create table community_authorization (
     id varchar primary key default uuid_generate_v4(),
     name varchar not null,
     api_key varchar not null,
-    max_surface double check (max_surface >= 0) not null
+    max_surface double precision check (max_surface >= 0) not null
 );
